requirements.txt
torch>=2.0.0
transformers>=4.35.0
torch-audio
accelerate
bitsandbytes
peft
datasets
pillow
opencv-python
easyocr
paddleocr
numpy
scipy
scikit-learn
matplotlib
seaborn
tqdm
wandb
python-dotenv
nltk
sentence-transformers
faiss-cpu
google-generativeai
